requirements.txt
%%writefile requirements.txt
streamlit
tensorflow
Pillow
numpy
